Art Uploads
Added sprite sheets for santa, and elves and pngs for the sleigh (wip) and the bg. Modified the parallax to loop with background at 5.75 scale and modified the player scene to include Santa's animation sprite with sprite sheets (not scripted yet) moved shooting point to hand location of attack and modified collision shape to match new sprite.